Reduce code duplication in ff (#7715)
* move distance constraint to forcefield

* move position and torsion constraint to force field, also remove some unused includes

* sorting includes

* add missing files
